Update JS-DOC and test naming
