+++ packages/protocol/src/action-protocol/element-hover.ts
@@ -34,9 +34,7 @@ export interface RequestPopupModelAction extends RequestAction<SetPopupModelActi
     elementId: string;
 
     /**
-     * The popup bounds declaring where the position where the popup should be rendered. Optionally the desired dimension
-     * can be set. If width & height are set to "-1" the client will auto compute the dimension when receiving the popup model
-     * from the server
+     * The popup bounds declaring the position of the popup. Optionally the desired dimension.
      */
     bounds: Bounds;
 }
