+++ packages/protocol/src/utils/type-util.spec.ts
@@ -15,67 +15,95 @@
  ********************************************************************************/
 import { expect } from 'chai';
 import { AnyObject, hasArrayProp, hasBooleanProp, hasNumberProp, hasObjectProp, hasStringProp } from './type-util';
+
 describe('TypeUtil', () => {
     describe('AnyObject', () => {
-        it('AnyObject.is with empty object', () => {
-            expect(AnyObject.is({})).to.be.true;
+        describe('is', () => {
+            it('should return true for an empty object', () => {
+                expect(AnyObject.is({})).to.be.true;
+            });
+            it('should return true for an object with arbitrary properties', () => {
+                expect(AnyObject.is({ a: 'a', b: 5 })).to.be.true;
+            });
+            it('should return false for `undefined`', () => {
+                expect(AnyObject.is(undefined)).to.be.false;
+            });
+            it('should return false for `null`', () => {
+                // eslint-disable-next-line no-null/no-null
+                expect(AnyObject.is(null)).to.be.false;
+            });
+            it('should return false for any object of primitive type', () => {
+                expect(AnyObject.is('')).to.be.false;
+                expect(AnyObject.is(5)).to.be.false;
+                expect(AnyObject.is(true)).to.be.false;
+
+                expect(
+                    AnyObject.is(() => {
+                        /**/
+                    })
+                ).to.be.false;
+                expect(AnyObject.is(Symbol('Symbol')));
+            });
         });
-        it('AnyObject.is with valid object', () => {
-            expect(AnyObject.is({ a: 'a', b: 5 })).to.be.true;
-        });
-        it('AnyObject.is with undefined', () => {
-            expect(AnyObject.is(undefined)).to.be.false;
-        });
-        it('AnyObject.is with null', () => {
-            // eslint-disable-next-line no-null/no-null
-            expect(AnyObject.is(null)).to.be.false;
-        });
-        it('AnyObject.is with non object types', () => {
-            expect(AnyObject.is('')).to.be.false;
-            expect(AnyObject.is(5)).to.be.false;
-            expect(AnyObject.is(true)).to.be.false;
+    });
 
-            expect(
-                AnyObject.is(() => {
-                    /**/
-                })
-            ).to.be.false;
-            expect(AnyObject.is(Symbol('Symbol')));
+    describe('hasStringProp', () => {
+        it('should return true for an object that has a property that matches the given key and type', () => {
+            expect(hasStringProp({ someProp: 'someKey' }, 'someProp')).to.be.true;
+        });
+        it('should return false for an object that has a property that matches the given but not the given type', () => {
+            expect(hasStringProp({ someProp: 123 }, 'someProp')).to.be.false;
+        });
+        it('should return false for an object that does not have a property that matches the given key.', () => {
+            expect(hasStringProp({ anotherProp: 123 }, 'someProp')).to.be.false;
         });
     });
 
-    it('hasStringProp with object with matching property', () => {
-        expect(hasStringProp({ someProp: 'someKey' }, 'someProp')).to.be.true;
-    });
-    it('hasStringProp with matching property but wrong type', () => {
-        expect(hasStringProp({ someProp: 123 }, 'someProp')).to.be.false;
-    });
-
-    it('hasBooleanProp with object with matching property', () => {
-        expect(hasBooleanProp({ someProp: true }, 'someProp')).to.be.true;
-    });
-    it('hasBooleanProp with matching property but wrong type', () => {
-        expect(hasBooleanProp({ someProp: undefined }, 'someProp')).to.be.false;
+    describe('hasBooleanProp', () => {
+        it('should return true for an object that has a property that matches the given key and type', () => {
+            expect(hasBooleanProp({ someProp: true }, 'someProp')).to.be.true;
+        });
+        it('should return false for an object that has a property that matches the given but not the given type', () => {
+            expect(hasBooleanProp({ someProp: 123 }, 'someProp')).to.be.false;
+        });
+        it('should return false for an object that does not have a property that matches the given key.', () => {
+            expect(hasBooleanProp({ anotherProp: 123 }, 'someProp')).to.be.false;
+        });
     });
 
-    it('hasNumberProp with object with matching property', () => {
-        expect(hasNumberProp({ someProp: 123 }, 'someProp')).to.be.true;
-    });
-    it('hasNumberProp with matching property but wrong type', () => {
-        expect(hasNumberProp({ someProp: '123' }, 'someProp')).to.be.false;
-    });
-
-    it('hasObjectProp with object with matching property', () => {
-        expect(hasObjectProp({ someProp: { value: 'someKey' } }, 'someProp')).to.be.true;
-    });
-    it('hasObjectProp with matching property but wrong type', () => {
-        expect(hasObjectProp({ someProp: 123 }, 'someProp')).to.be.false;
+    describe('hasNumberProp', () => {
+        it('should return true for an object that has a property that matches the given key and type', () => {
+            expect(hasNumberProp({ someProp: 123 }, 'someProp')).to.be.true;
+        });
+        it('should return false for an object that has a property that matches the given but not the given type', () => {
+            expect(hasNumberProp({ someProp: '123' }, 'someProp')).to.be.false;
+        });
+        it('should return false for an object that does not have a property that matches the given key.', () => {
+            expect(hasBooleanProp({ anotherProp: 123 }, 'someProp')).to.be.false;
+        });
     });
 
-    it('hasArrayProp with object with matching property', () => {
-        expect(hasArrayProp({ someProp: ['some', 'prop'] }, 'someProp')).to.be.true;
+    describe('hasObjectProp', () => {
+        it('should return true for an object that has a property that matches the given key and type', () => {
+            expect(hasObjectProp({ someProp: { value: 'someKey' } }, 'someProp')).to.be.true;
+        });
+        it('should return false for an object that has a property that matches the given but not the given type', () => {
+            expect(hasObjectProp({ someProp: '123' }, 'someProp')).to.be.false;
+        });
+        it('should return false for an object that does not have a property that matches the given key.', () => {
+            expect(hasObjectProp({ anotherProp: 123 }, 'someProp')).to.be.false;
+        });
     });
-    it('hasArrayProp with matching property but wrong type', () => {
-        expect(hasArrayProp({ someProp: 123 }, 'someProp')).to.be.false;
+
+    describe('hasArrayProp', () => {
+        it('should return true for an object that has a property that matches the given key and type', () => {
+            expect(hasArrayProp({ someProp: ['some', 'prop'] }, 'someProp')).to.be.true;
+        });
+        it('should return false for an object that has a property that matches the given but not the given type', () => {
+            expect(hasArrayProp({ someProp: '123' }, 'someProp')).to.be.false;
+        });
+        it('should return false for an object that does not have a property that matches the given key.', () => {
+            expect(hasArrayProp({ anotherProp: 123 }, 'someProp')).to.be.false;
+        });
     });
 });
